Pull gatherLoopDetails into its own file
